--- code-review/index.ejs
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <title>Martian</title>
-  <link rel='stylesheet' href='/css/reset.css' />
-  <link rel='stylesheet' href='/css/style.css' />
-</head>
-
-<body>
-  <canvas class="wheel" width="400" height="400"></canvas>
-
-
-  <div class="main__wrapper">
-    <div class="main__receive">
-      <div>송수신정보</div>
-      <input type="text">
-      <button>해석하기</button>
-    </div>
-    <div class="main__send">
-      <div>발신정보입력</div>
-      <input type="text" placeholder="> hello earth">
-      <button>지구로 보내기</button>
-    </div>
-  </div>
-
-  <script type="module" src="./js/main.js"></script>
-</body>
-
-</html>
