+++ martian/public/js/Communicator.js
@@ -1,0 +1,1 @@
+import { _ } from './util.js'
