+++ martian/views/index.ejs
@@ -8,7 +8,31 @@
 </head>
 
 <body>
-  <canvas class="wheel" width="400" height="400"></canvas>
+  <div class="wheel">
+    <canvas class="wheel__circle" width="400" height="400"></canvas>
+    <div class="wheel__letter">
+      <span class="letter">5</span>
+      <span class="letter">6</span>
+      <span class="letter">7</span>
+      <span class="letter">8</span>
+      <span class="letter">9</span>
+      <span class="letter">A</span>
+      <span class="letter">B</span>
+      <span class="letter">C</span>
+      <span class="letter">D</span>
+      <span class="letter">E</span>
+      <span class="letter">F</span>
+      <span class="letter">0</span>
+      <span class="letter">1</span>
+      <span class="letter">2</span>
+      <span class="letter">3</span>
+      <span class="letter">4</span>
+    </div>
+    <div class="wheel__arrow">
+      <img src="./images/arrow.png" alt="">
+    </div>
+  </div>
+
 
 
   <div class="main__wrapper">
